move MockedRfqtFirmQuoteResponse into types file
